+++ java-sdk/src/main/java/com/netflix/conductor/sdk/workflow/executor/WorkflowExecutor.java
@@ -22,6 +22,7 @@ import java.util.concurrent.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.netflix.conductor.client.exception.ConductorClientException;
 import com.netflix.conductor.client.http.MetadataClient;
 import com.netflix.conductor.client.http.TaskClient;
 import com.netflix.conductor.client.http.WorkflowClient;
@@ -162,10 +163,15 @@ public class WorkflowExecutor {
         request.setName(name);
         request.setVersion(version);
 
-        String workflowId = workflowClient.startWorkflow(request);
-        runningWorkflowFutures.put(workflowId, future);
-
-        return future;
+        try {
+            String workflowId = workflowClient.startWorkflow(request);
+            runningWorkflowFutures.put(workflowId, future);
+            return future;
+        } catch (ConductorClientException cce) {
+            cce.printStackTrace();
+            System.out.println("code: " + cce.getStatus());
+        }
+        return new CompletableFuture<>();
     }
 
     public CompletableFuture<Workflow> executeWorkflow(
