+++ java-sdk/src/test/java/com/netflix/conductor/sdk/workflow/def/TaskConversionsTests.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.netflix.conductor.sdk.workflow.def.tasks;
+package com.netflix.conductor.sdk.workflow.def;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -21,12 +21,12 @@ import org.junit.jupiter.api.Test;
 
 import com.netflix.conductor.common.metadata.workflow.WorkflowTask;
 import com.netflix.conductor.common.run.Workflow;
-import com.netflix.conductor.sdk.workflow.def.ValidationError;
+import com.netflix.conductor.sdk.workflow.def.tasks.*;
 import com.netflix.conductor.sdk.workflow.executor.WorkflowExecutor;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class TestTaskConversions {
+public class TaskConversionsTests {
 
     static {
         WorkflowExecutor.initTaskImplementations();
