+++ java-sdk/src/test/java/com/netflix/conductor/sdk/workflow/testing/Task1Input.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Netflix, Inc.
+ * Copyright 2021 Netflix, Inc.
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -10,27 +10,32 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.netflix.conductor.sdk;
+package com.netflix.conductor.sdk.workflow.testing;
 
-public class MyWorkflowState {
+public class Task1Input {
 
-    private boolean paymentCompleted;
+    private int mod;
 
-    private int timeTaken;
+    private int oddEven;
 
-    public boolean isPaymentCompleted() {
-        return paymentCompleted;
+    public int getMod() {
+        return mod;
     }
 
-    public void setPaymentCompleted(boolean paymentCompleted) {
-        this.paymentCompleted = paymentCompleted;
+    public void setMod(int mod) {
+        this.mod = mod;
     }
 
-    public int getTimeTaken() {
-        return timeTaken;
+    public int getOddEven() {
+        return oddEven;
     }
 
-    public void setTimeTaken(int timeTaken) {
-        this.timeTaken = timeTaken;
+    public void setOddEven(int oddEven) {
+        this.oddEven = oddEven;
+    }
+
+    @Override
+    public String toString() {
+        return "Task1Input{" + "mod=" + mod + ", oddEven=" + oddEven + '}';
     }
 }
