--- java-sdk/src/test/java/com/netflix/conductor/tests/TestWorkflowInput.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2022 Netflix, Inc.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package com.netflix.conductor.tests;
-
-public class TestWorkflowInput {
-
-    private String name;
-
-    private String zipCode;
-
-    private String countryCode;
-
-    public TestWorkflowInput(String name, String zipCode, String countryCode) {
-        this.name = name;
-        this.zipCode = zipCode;
-        this.countryCode = countryCode;
-    }
-
-    public TestWorkflowInput() {}
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getZipCode() {
-        return zipCode;
-    }
-
-    public void setZipCode(String zipCode) {
-        this.zipCode = zipCode;
-    }
-
-    public String getCountryCode() {
-        return countryCode;
-    }
-
-    public void setCountryCode(String countryCode) {
-        this.countryCode = countryCode;
-    }
-}
